Add examples for fine-tuning to examples folder
