+++ examples/fine_tune_models/configs/geneformer_config.yaml
@@ -1,0 +1,6 @@
+model_name: "gf-12L-30M-i2048"
+batch_size: 24
+emb_layer: -1
+emb_mode: "cell"
+device: "cpu"
+accelerator: False
