+++ examples/fine_tune_models/configs/uce_config.yaml
@@ -1,0 +1,18 @@
+model_name: "4layer_model"
+batch_size: 5
+species : "human"
+gene_embedding_model: "ESM2"
+pad_length: 1536
+pad_token_idx: 0
+chrom_token_left_idx: 1
+chrom_token_right_idx: 2
+cls_token_idx: 3
+CHROM_TOKEN_OFFSET: 143574
+sample_size: 1024
+CXG: True
+output_dim: 1280
+d_hid: 5120
+token_dim: 5120
+multi_gpu: False
+device: "cpu"
+accelerator: False
