+++ helical/models/uce/fine_tuning_model.py
@@ -1,8 +1,19 @@
+from helical.models.fine_tune.fine_tuning_heads import ClassificationHead
+from helical.models.uce.uce_dataset import UCEDataset
 import torch
+from torch import optim
+from torch.nn.modules import loss
+from torch.utils.data import DataLoader
 from helical.models.base_models import HelicalBaseFineTuningHead, HelicalRNAModel
+from helical.models.base_models import HelicalBaseFineTuningModel
+from sklearn.metrics import accuracy_score
+from typing import Literal, Optional
+from tqdm import tqdm
+from transformers import get_scheduler
 
-class UCEFineTuningModel(torch.nn.Module):
-    """Fine-tuning model for the UCE model.
+class UCEFineTuningModel(HelicalBaseFineTuningModel):
+    """
+    Fine-tuning model for the UCE model.
 
     Parameters
     ----------
@@ -10,25 +21,171 @@ class UCEFineTuningModel(torch.nn.Module):
         The initialised UCE model to fine-tune.
     fine_tuning_head : HelicalBaseFineTuningHead
         The fine-tuning head that is appended to the model.
+    output_size : Optional[int]
+        The output size of the fine-tuning model. This is required if the fine_tuning_head is a string specified task. For a classification task this is number of unique classes.
 
     Methods
     -------
     forward(input_gene_ids: torch.Tensor, data_dict: dict, src_key_padding_mask: torch.Tensor, use_batch_labels: bool, device: str) -> torch.Tensor
         The forward method of the fine-tuning model.
     """
-    def __init__(self, uce_model: HelicalRNAModel, fine_tuning_head: HelicalBaseFineTuningHead):
+    def __init__(self, uce_model: HelicalRNAModel, fine_tuning_head: Literal["classification"]|HelicalBaseFineTuningHead, output_size: Optional[int]=None):
         super(UCEFineTuningModel, self).__init__()
+        self.config = uce_model.config
         self.helical_model = uce_model.model
+        self.device = uce_model.device
+        self.accelerator = uce_model.accelerator
+        if isinstance(fine_tuning_head, str):
+            if fine_tuning_head == "classification":
+                if output_size is None:
+                    raise ValueError("The output_size must be specified for a classification head.")
+                fine_tuning_head = ClassificationHead(output_size)
+            else:
+                raise ValueError(f"The fine_tuning_head must be a valid HelicalBaseFineTuningHead")
+        else:
+            fine_tuning_head = fine_tuning_head
+        fine_tuning_head.set_dim_size(self.config["embsize"])
         self.fine_tuning_head = fine_tuning_head
 
     def forward(self, batch_sentences, mask) -> torch.Tensor:
         _, embeddings = self.helical_model.forward(batch_sentences, mask=mask)
-        # if self.helical_model.accelerator is not None:
-        #     self.accelerator.wait_for_everyone()
-        #     embeddings = self.helical_model.accelerator.gather_for_metrics((embedding))
-        #     if self.helical_model.accelerator.is_main_process:
-        #         embeddings = embedding
-        # else:
-        #     embeddings = embedding
+        if self.accelerator is not None:
+            self.accelerator.wait_for_everyone()
+            embeddings = self.accelerator.gather_for_metrics((embeddings))
+            if self.accelerator.is_main_process:
+                embeddings = embeddings
+        else:
+            embeddings = embeddings
         output = self.fine_tuning_head(embeddings)
         return output
+    
+    def fine_tune(
+            self,
+            train_input_data: UCEDataset, 
+            train_labels,     
+            validation_input_data = None,
+            validation_labels = None,
+            optimizer: optim = optim.AdamW,
+            optimizer_params: dict = {'lr': 0.0001}, 
+            loss_function: loss = loss.CrossEntropyLoss(), 
+            epochs: int = 1,
+            # freeze_layers: int = 0,
+            lr_scheduler_params: Optional[dict] = None):
+        """
+        Fine-tunes the UCE model with different head modules. 
+
+        Parameters
+        ----------
+        train_input_data : Dataset
+            A helical UCE processed dataset for fine-tuning
+        train_labels : ndarray
+            The labels for the training data. These should be stored as unique per class integers.
+        validation_input_data : Dataset, default = None
+            A helical UCE processed dataset for per epoch validation. If this is not specified, no validation will be performed.
+        validation_labels : ndarray, default = None,
+            The labels for the validation data. These should be stored as unique per class integers.
+        optimizer : torch.optim, default = torch.optim.AdamW
+            The optimizer to be used for training.
+        optimizer_params : dict
+            The optimizer parameters to be used for the optimizer specified. This list should NOT include model parameters.
+            e.g. optimizer_params = {'lr': 0.0001}
+        loss_function : torch.nn.modules.loss, default = torch.nn.modules.loss.CrossEntropyLoss()
+            The loss function to be used.
+        epochs : int, optional, default = 10
+            The number of epochs to train the model
+        freeze_layers : int, optional, default = 0
+            The number of layers to freeze.
+        lr_scheduler_params : dict, default = None
+            The learning rate scheduler parameters for the transformers get_scheduler method. The optimizer will be taken from the optimizer input and should not be included in the learning scheduler parameters. If not specified, no scheduler will be used.
+            e.g. lr_scheduler_params = { 'name': 'linear', 'num_warmup_steps': 0, 'num_training_steps': 5 }
+
+        Returns
+        -------
+        torch.nn.Module
+            The fine-tuned model.
+        """
+        batch_size = self.config["batch_size"]
+        dataloader = DataLoader(train_input_data, 
+                                batch_size=batch_size, 
+                                shuffle=False,
+                                collate_fn=train_input_data.collator_fn,
+                                num_workers=0)
+        
+        if validation_input_data is not None:
+            validation_dataloader = DataLoader(validation_input_data, 
+                        batch_size=batch_size, 
+                        shuffle=False,
+                        collate_fn=validation_input_data.collator_fn,
+                        num_workers=0)
+
+        if self.accelerator is not None:
+            dataloader = self.accelerator.prepare(dataloader)
+            if validation_input_data is not None:
+                validation_dataloader = self.accelerator.prepare(validation_dataloader)
+
+
+        # disable progress bar if not the main process
+        # if self.accelerator is not None:
+        #     pbar = tqdm(dataloader, disable=not self.accelerator.is_local_main_process)
+        # else:
+        #     pbar = tqdm(dataloader)
+
+        model = self.to(self.device)
+        
+        model.train()
+
+        optimizer = optimizer(model.parameters(), **optimizer_params)
+
+        lr_scheduler = None
+        if lr_scheduler_params is not None: 
+            lr_scheduler = get_scheduler(optimizer=optimizer, **lr_scheduler_params)
+
+        for j in range(epochs):
+            batch_count = 0
+            batch_loss = 0.0
+            batches_processed = 0
+            training_loop = tqdm(dataloader, desc="Fine-Tuning")
+            for batch in training_loop:
+                batch_sentences, mask, idxs = batch[0], batch[1], batch[2]
+                batch_sentences = batch_sentences.permute(1, 0)
+                if self.config["multi_gpu"]:
+                    batch_sentences = self.helical_model.module.pe_embedding(batch_sentences.long())
+                else:
+                    batch_sentences = self.helical_model.pe_embedding(batch_sentences.long())
+                batch_sentences = torch.nn.functional.normalize(batch_sentences, dim=2)  # normalize token outputs
+                output = model.forward(batch_sentences, mask=mask)
+                labels = torch.tensor(train_labels[batch_count: batch_count + self.config["batch_size"]], device=self.device)
+                batch_count += self.config["batch_size"]
+                loss = loss_function(output, labels)
+                loss.backward()
+                batch_loss += loss.item()
+                batches_processed += 1
+
+                optimizer.step()
+                optimizer.zero_grad()
+
+                training_loop.set_postfix({"loss": batch_loss/batches_processed})
+                training_loop.set_description(f"Fine-Tuning: epoch {j+1}/{epochs}")
+
+            if lr_scheduler is not None:
+                lr_scheduler.step()
+
+            if validation_input_data is not None:
+                testing_loop = tqdm(validation_dataloader, desc="Fine-Tuning Validation")
+                accuracy = 0.0
+                count = 0.0
+                validation_batch_count = 0
+                for validation_data in testing_loop:
+                    batch_sentences, mask, idxs = validation_data[0], validation_data[1], validation_data[2]
+                    batch_sentences = batch_sentences.permute(1, 0)
+                    if self.config["multi_gpu"]:
+                        batch_sentences = self.helical_model.module.pe_embedding(batch_sentences.long())
+                    else:
+                        batch_sentences = self.helical_model.pe_embedding(batch_sentences.long())
+                    batch_sentences = torch.nn.functional.normalize(batch_sentences, dim=2)  # normalize token outputs
+                    output = model.forward(batch_sentences, mask=mask)
+                    val_labels = torch.tensor(validation_labels[validation_batch_count: validation_batch_count + self.config["batch_size"]], device=self.device)
+                    validation_batch_count += self.config["batch_size"]
+                    accuracy += accuracy_score(val_labels.cpu(), torch.argmax(output, dim=1).cpu())
+                    count += 1.0
+                    testing_loop.set_postfix({"accuracy": accuracy/count})
